refactor: extract sandbox reified tests
